Merged PR 858519: Move inline-text-overflow to normalize.scss
We can't define classes in _constants.scss because it gets included in pretty much every scss and gets mangled, so we end up with a ton of duplicate declarations of these classes. normalize.scss gets included only once (by Shell), so place common classes there and mark them as `:global` so the names aren't mangled.